+++ src/components/interaction/loading-indicator/loading-indicator.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="flex items-center gap-2" aria-live="polite" data-test="loading-indicator">
-		<icon-loading class="spin text-grey-500" />
+		<icon-loading class="spin text-grey-500 dark:text-grey-400" />
 		<slot />
 	</div>
 </template>