+++ src/components/form/radio-group/radio-group.vue
@@ -1,6 +1,6 @@
 <template>
-	<fieldset class="flex flex-col gap-1" v-bind="{ 'aria-describedby': describedBy }" data-test="radio-group">
-		<form-label v-bind="{ tag: 'legend' }" class="mb-2">
+	<fieldset class="flex flex-col gap-2 @container" v-bind="{ 'aria-describedby': describedBy }" data-test="radio-group">
+		<form-label v-bind="{ tag: 'legend' }" class="mb-1">
 			<slot />
 		</form-label>
 
@@ -9,7 +9,7 @@
 		</conditional-wrapper>
 
 		<slot name="options" v-bind="{ options: internalOptions, name: name || inputId }">
-			<div class="flex" :class="{ 'gap-10': inline, 'flex-col gap-2': !inline }">
+			<div class="flex flex-col" :class="{ '@xs:flex-row @xs:gap-10': inline, 'gap-2': !inline }">
 				<template v-for="option in internalOptions" :key="option.id">
 					<div class="flex items-center gap-2">
 						<input v-model="model" type="radio" class="form-radio" v-bind="{ id: option.id, value: option.value, name: name || inputId }" />